Multiqueries for typetoken_lemma and queryscripts for typetoken_bedeutung added
-multiqueries for typetoken_lemma and typetoken_lemma_name queries added
-scripts for query typetoken_bedeutung added
-query_all.sh and README updated

+++ 04_query_db/03_typetoken_lemma/db_query_typetoken_lemma.sh
@@ -22,14 +22,24 @@
 DB=../../../02_generated_files/corpus.db
 
 if [ -z $1 ] || [ -z $2 ] || [ -z $3 ] || [ -z $4 ] || [ ! -d $4 ] ; then
-	echo "usage $0 daymin daymax postag outputdir"
+	echo "usage $0 daymin daymax postag outputdir (hint: postag can also be a list like AAA+BBB...)"
 	exit
 fi
 
 DAYMIN=$1
 DAYMAX=$2
-POSTAG="$3"
+POSTAGLIST="$3"
 OUTPUTDIR=$4
+
+# build the query for postag or postag-list
+POSTAG_QUERY="AND ( "
+for i in $(echo $POSTAGLIST | tr "+" "\n") ; do
+	POSTAG_QUERY=${POSTAG_QUERY}'textpostags.postag = "'
+	POSTAG_QUERY=${POSTAG_QUERY}${i}
+	POSTAG_QUERY=${POSTAG_QUERY}'" '
+	POSTAG_QUERY=${POSTAG_QUERY}"OR "
+done
+POSTAG_QUERY=${POSTAG_QUERY}'0 ) '
 
 #echo "--------------------------- $POSTAG ----------------------------"
 echo -n 'SELECT textpostags.lemma ' > query.sql
@@ -40,46 +50,50 @@ echo -n $DAYMIN >> query.sql
 echo -n ' AND julianday(aufnahme.datum) ' >> query.sql
 echo -n '- julianday(sprecher.geburtsdatum) < ' >> query.sql
 echo -n $DAYMAX >> query.sql
-echo -n ' AND textpostags.postag = "' >> query.sql
-echo -n $POSTAG >> query.sql
-echo -n '" AND aufnahme.sprecher_id = sprecher.id ' >> query.sql
+echo -n ' ' >> query.sql
+echo -n $POSTAG_QUERY >> query.sql
+echo -n ' AND aufnahme.sprecher_id = sprecher.id ' >> query.sql
 echo -n 'AND textpostags.aufnahme_id = aufnahme.id;' >> query.sql
 
-#cat query.sql | sqlite3 $DB | wc -l
-cat query.sql | sqlite3 $DB | sort | uniq > \
-		$OUTPUTDIR/output_wortlist_${POSTAG}_${DAYMIN}_${DAYMAX}.txt
-WORTLIST=`cat $OUTPUTDIR/output_wortlist_${POSTAG}_${DAYMIN}_${DAYMAX}.txt`
-CNT=0
-WORDCNTSUM=0
+COUNT=`cat query.sql | sqlite3 $DB | wc -l`
+echo ", $COUNT entries found"
+if [ $COUNT -gt 0 ] ; then
+	cat query.sql | sqlite3 $DB | sort | uniq > \
+			$OUTPUTDIR/output_wortlist_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.txt
+	WORTLIST=`cat $OUTPUTDIR/output_wortlist_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.txt`
+	CNT=0
+	WORDCNTSUM=0
 
-rm -f $OUTPUTDIR/output_typetoken_lemma_${POSTAG}_${DAYMIN}_${DAYMAX}.csv
-for j in $WORTLIST ; do
-	#echo "........................... $j ............................"
-	echo -n 'SELECT textpostags.postag ' > query2.sql
-	echo -n 'FROM sprecher,aufnahme,textpostags ' >> query2.sql
-	echo -n 'WHERE julianday(aufnahme.datum) ' >> query2.sql
-	echo -n '- julianday(sprecher.geburtsdatum) > ' >> query2.sql
-	echo -n $DAYMIN >> query2.sql
-	echo -n ' AND julianday(aufnahme.datum) ' >> query2.sql
-	echo -n '- julianday(sprecher.geburtsdatum) < ' >> query2.sql
-	echo -n $DAYMAX >> query2.sql
-	echo -n ' AND textpostags.postag = "' >> query2.sql
-	echo -n $POSTAG >> query2.sql
-	echo -n '" AND textpostags.lemma = '\' >> query2.sql
-	echo -n $j\' >> query2.sql
-	echo -n ' AND aufnahme.sprecher_id = sprecher.id ' >> query2.sql
-	echo -n 'AND textpostags.aufnahme_id = aufnahme.id;' >> query2.sql
-	#cat query2.sql | sqlite3 $DB
-	echo -n "$j|" \
-	>> $OUTPUTDIR/output_typetoken_lemma_${POSTAG}_${DAYMIN}_${DAYMAX}.csv
-	WORDCNT=`cat query2.sql | sqlite3 $DB | wc -l`
-	WORDCNTSUM=$((WORDCNTSUM + WORDCNT))
-	CNT=$((CNT + 1))
-	echo $WORDCNT >> $OUTPUTDIR/output_typetoken_lemma_${POSTAG}_${DAYMIN}_${DAYMAX}.csv
-done
-echo "----|----" >> $OUTPUTDIR/output_typetoken_lemma_${POSTAG}_${DAYMIN}_${DAYMAX}.csv
-echo "$CNT|$WORDCNTSUM" >> $OUTPUTDIR/output_typetoken_lemma_${POSTAG}_${DAYMIN}_${DAYMAX}.csv
-echo "${POSTAG}|${DAYMIN}|${DAYMAX}||$CNT|$WORDCNTSUM" >> $OUTPUTDIR/../output_typetoken_lemma_summary.csv
+	rm -f $OUTPUTDIR/output_typetoken_lemma_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.csv
+	for j in $WORTLIST ; do
+		#echo "........................... $j ............................"
+		echo -n 'SELECT textpostags.postag ' > query2.sql
+		echo -n 'FROM sprecher,aufnahme,textpostags ' >> query2.sql
+		echo -n 'WHERE julianday(aufnahme.datum) ' >> query2.sql
+		echo -n '- julianday(sprecher.geburtsdatum) > ' >> query2.sql
+		echo -n $DAYMIN >> query2.sql
+		echo -n ' AND julianday(aufnahme.datum) ' >> query2.sql
+		echo -n '- julianday(sprecher.geburtsdatum) < ' >> query2.sql
+		echo -n $DAYMAX >> query2.sql
+		echo -n ' ' >> query2.sql
+		echo -n $POSTAG_QUERY >> query2.sql
+		echo -n ' AND textpostags.lemma = '\' >> query2.sql
+		echo -n $j\' >> query2.sql
+		echo -n ' AND aufnahme.sprecher_id = sprecher.id ' >> query2.sql
+		echo -n 'AND textpostags.aufnahme_id = aufnahme.id;' >> query2.sql
+		#cat query2.sql | sqlite3 $DB
+		echo -n "$j|" \
+		>> $OUTPUTDIR/output_typetoken_lemma_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.csv
+		WORDCNT=`cat query2.sql | sqlite3 $DB | wc -l`
+		WORDCNTSUM=$((WORDCNTSUM + WORDCNT))
+		CNT=$((CNT + 1))
+		echo $WORDCNT >> $OUTPUTDIR/output_typetoken_lemma_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.csv
+	done
+	echo "----|----" >> $OUTPUTDIR/output_typetoken_lemma_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.csv
+	echo "$CNT|$WORDCNTSUM" >> $OUTPUTDIR/output_typetoken_lemma_${POSTAGLIST}_${DAYMIN}_${DAYMAX}.csv
+	echo "${POSTAGLIST}|${DAYMIN}|${DAYMAX}||$CNT|$WORDCNTSUM" >> $OUTPUTDIR/../output_typetoken_lemma_summary_${DAYMIN}_${DAYMAX}.csv
 
+	rm -f query2.sql
+fi
 rm -f query.sql
-rm -f query2.sql
+
